+++ deeprankcore/features/contact.py
@@ -7,7 +7,7 @@ from deeprankcore.models.structure.atom import Atom
 from deeprankcore.models.graph import Graph
 from deeprankcore.models.contact import ResidueContact, AtomicContact
 from deeprankcore.domain import edgefeatures as Efeat
-from deeprankcore.domain.forcefield.parsers import atomic_forcefield
+from deeprankcore.models.operations.parsers import atomic_forcefield
 
 _log = logging.getLogger(__name__)
