+++ tests/domain/test_aminoacids.py
