+++ tests/domain/test_forcefield.py
@@ -1,5 +1,5 @@
 from pdb2sql import pdb2sql
-from deeprankcore.domain.forcefield.parsers import atomic_forcefield
+from deeprankcore.models.operations.parsers import atomic_forcefield
 from deeprankcore.models.operations.pdb import get_structure
 from deeprankcore.domain.aminoacids import arginine, glutamate
