+++ tests/models/test_community_pooling.py
